chore: Balance Bar Fragment
