+++ app/src/main/res/layout/activity_home.xml
@@ -109,58 +109,12 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/balanceBarContainer"
-        android:padding="20dp"
+    <FrameLayout
+        android:id="@+id/balanceBarFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="20dp"
-        android:background="@color/background_light"
-        app:layout_constraintTop_toBottomOf="@id/balanceContainer">
-
-        <View
-            android:id="@+id/negative"
-            android:layout_width="match_parent"
-            android:layout_height="15dp"
-            android:background="@drawable/light_black_corner"
-            android:backgroundTint="@color/red"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.99"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <View
-            android:id="@+id/positive"
-            android:layout_width="match_parent"
-            android:layout_height="15dp"
-            android:background="@drawable/light_black_corner"
-            android:backgroundTint="@color/text_secondary"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/positiveValue"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="$ 0.00"
-            app:layout_constraintTop_toBottomOf="@id/positive"
-            app:layout_constraintStart_toStartOf="@id/positive"
-            android:textColor="@color/text_secondary"
-            android:textSize="14sp" />
-
-        <TextView
-            android:id="@+id/negativeValue"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/positive"
-            app:layout_constraintEnd_toEndOf="@id/negative"
-            android:text="$ 0.00"
-            android:textColor="@color/text_secondary"
-            android:textSize="14sp"/>
-    </androidx.constraintlayout.widget.ConstraintLayout>
+        app:layout_constraintTop_toBottomOf="@id/balanceContainer"
+        app:layout_constraintStart_toStartOf="parent"/>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/footer"
